genesis: use non-zero timestamps
as suggested

Ceyhun Onur:
"warp cannot be activated at timestamp = 0
"which is before Durango
"also set genesis timestamp  to some non-zero value (a recent timestamp is fine).

Gene Aumson:
"thanks. just to confirm, is there a recent change that started enforcing this? again, this genesis file was working not long ago"

Ceyhun Onur:
"yes it was changed after v0.6.3 when we introduced modifiable network upgrades and fixing a regression with using 0 timestamps in genesis."